[CL 5/5] NAI4FL code model: code model v0 for localization of faults in UAST programs
[Cloned from CL 237369945 by 'hg patch'.
Original change by dbieber@dbieber:z2019-03-07-code-model-cls:3145:citc on 2019/03/07 18:35:44.]

Introduces the code localization model.

A program consists of ~30 statements, each of which consists of ~10 tokens.
Such a program has about 25 components. This number can vary highly as the meaning of component changes between  fault injectors.

In this model (the "code model"), we embed each statement using an LSTM over the token embeddings.
And then we embed the program using an LSTM over the statement embeddings.
We use this program embedding as a proxy for the embedding of the distribution of outputs the program could produce.

Contrast the code model with the unimplemented "trace model" and "control flow model" planned for future CLs.

See cl/236175012 for this CL in the context of the CLs leading up to this change.

Notes:
The current model handles about 50 examples (5 batches of 10 examples) in 80 minutes on a CPU. That's too slow!

----

Contents:
1. cl/237470192: component aware uast tokenization library
2. cl/237470182: New feature connector -- code_tokens_feature.py
3. cl/237470177: fault injectors API update
4. cl/237470188: updates to NAPS dataset in tfds
5. cl/237470173: [this CL] code model for localization of faults
PiperOrigin-RevId: 237470173