[CL 1/5] NAI4FL code model: adds component aware uast tokenization library.
[Cloned from CL 237370017 by 'hg patch'.
Original change by dbieber@dbieber:z2019-03-07-code-model-cls:3145:citc on 2019/03/07 18:36:43.]

See cl/236175012 for the context of how the uast_tokenization gets used.

This library provides component-aware tokenization of NAPS UASTs. A program is
made of many statements, and a program also has many components. The meaning of
'component' varies between fault injection techniques.

When a program is tokenized, each statement is represented as a sequence of
tokens. Each token in the resulting tokenization has a string value as well
as an optional component_index. The component_index indicates which component,
if any, the token is a part of in the program.

The component indexes make it possible to mask out individual components from
tokenized programs.

Contents:
1. cl/237470192: [this CL] component aware uast tokenization library
2. cl/237470182: New feature connector -- code_tokens_feature.py
3. cl/237470177: fault injectors API update
4. cl/237470188: updates to NAPS dataset in tfds
5. cl/237470173: code model for localization of faults
PiperOrigin-RevId: 237470192