Expose sample_rate (in Hertz) metadata to audio datasets.
This is pretty critical metadata to know when designing (e.g. spectrogram) feature extraction parameters or simply listening to audio samples.  Exposing it in the DatasetInfo allows users to avoid having to look it up and specify it manually.

Had to fix the BeamMetadataDict constructor to pass args through to the base dict constructor in order to make this work.

PiperOrigin-RevId: 296499129