Add TED-LIUM release 3 speech recognition corpus.
Note that this only includes a config for the "legacy" splits for
TED-LIUM 3 for now to remain consistent with releases 1 and 2.

Had to work around a bug in the tarball which contains invalid symlinks.  These invalid symlinks are only in the non-legacy "speaker adaptation" splits, so supporting those splits will require a different workaround.

PiperOrigin-RevId: 307718887