More flexible "supervised_key" handling.
The current implementation only allows a you to create a `(feature, target)` pair.

This change expands the behavior to allow tuples of 2 or 3 elements (keras also recognizes `(features, targets, sample_weights)` tuples) containing arbitrary `tf.nest` style structures of tuple, dict, and string feature-keys.

This way if the canonical task for a dataset expects a more complex structure, you can define it here for keras. Typically this will still be a tuple, but this change allows you to have dictionaries of features in either the `features` or `targets` element of the tuple, without requiring that you nest the FeaturesDict objects for the main dataset.

For the titanic dataset this allows you to use a standard flat dictionary of features, so that "survived" is not special unless you set `as_supervised=True`

PiperOrigin-RevId: 392720807