Update `SupervisedKeySection.get_key` to support dict/tuples.
`build_submit_catalog` has been failing on the newly added `penguins.simple` dataset, which uses the newly added support for dictionaries in `supervised_keys`. As `dict`s are mutable & unhashable, they can't be used in a set (see `_display_all_builders`).

The markdown builder was using the value of supervised_keys to build a set, but the signature for `get_key` is `Key = Union[int, str]`. To match the type signature, I've used `repr` to convert the nested structures in a way that's hashable & comparable.

PiperOrigin-RevId: 395629598