First try to find a dataset without enforcing a default config.
When a dataset first was without configs, but later got changed into having configs, the main dataset folder will contain subfolders for configs and old versions. Before, the default config was always used to load a version and consequently it was impossible to load old versions without a config. This change first tries to find the dataset dir with the config as given and then after tries to find it with the default config.

PiperOrigin-RevId: 453396216