Simplify Beam pipeline by removing finalize method
Currently it is not possible to give TFDS dataset builder a pcollection. Namely, when TFDS is given a pcollection, then it will execute split_builder.submit_split_generation, which requires to have control over when the Beam pipeline is executed because it invokes the BeamWriter.finalize method, which assumes the pipeline is finished. However, with an external pcollection TFDS has no control over when the pipeline is executed.

PiperOrigin-RevId: 502871614