Lazy load array_record.
This is a short-term solution to the installation problems on macOS since 4.9.0.
This also makes sense in the long-term, because array_record is not used by all users.

The long-term solution to fix macOS installation is to build proper wheels for array_record on macos/windows platforms, but this will take more time.

PiperOrigin-RevId: 523373538