Yield found dataset references even when some exception was invoked
If any of the paths cannot be read by `register.list_dataset_references`, it yields all the found references and raises an exception at the end. Before this change an empty list of references was returned whenever there was a problem with a path, so you would miss all the references that were found.

PiperOrigin-RevId: 615421306